requirements.txt
asgiref==3.5.1
autopep8==1.7.0
black==22.3.0
certifi==2021.10.8
click==8.1.3
colorama==0.4.4
Cython==0.29.28
distlib==0.3.4
Django==3.2.13
django-extensions==3.1.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.7.0
flake8==4.0.1
gensim==4.2.0
graphviz==0.15
isort==5.10.1
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.21.6
pathspec==0.9.0
Pillow==9.1.1
pipenv==2022.5.2
platformdirs==2.5.2
pycodestyle==2.9.1
pydot==1.4.2
pyflakes==2.4.0
PyJWT==2.6.0
pyparsing==3.0.9
pytz==2022.1
scipy==1.7.3
six==1.16.0
smart-open==6.0.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tzdata==2022.1
uuid==1.30
virtualenv==20.14.1
virtualenv-clone==0.5.7
